+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [UNRELEASED]
+## Changed
+- Add `disableWrapperClick` props to `Modal` component
+
 ## [0.9.5] - 2019-03-19
 ## Changed
 - Add `themeColor` props to `Modal` component